People Web App
Built simple web app to practice basic CRUD operations, usage of Hibernate for communicating with database, and bootstrap for tables, buttons, and usage of spring framework for input validations

This app is for educational practice only and simply tracks a list of people with first and last name, date of birth, email, and salary. People are stored in a database, and can be deleted or updated, new people records can be created. All fields are required for new data records except for salary, email addresses must be valid, and dates of birth must be in the past.